implements editors
